requirements.txt
streamlit
requests
python-dotenv
google-generativeai
supabase
